// webapp/src/api/model/pnPUserCreation.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PnPUserCreationAuthoritiesItem } from './pnPUserCreationAuthoritiesItem';

export interface PnPUserCreation {
  authorities: PnPUserCreationAuthoritiesItem[];
  displayName: string;
  email?: string;
  password?: string;
  username: string;
}